+++ src/Attributes/CollectionResults/DoesntContainStrict.php
@@ -1,0 +1,39 @@
+<?php declare(strict_types=1);
+
+/**
+ * Copyright (C) Brian Faust
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cline\Struct\Attributes\CollectionResults;
+
+use Attribute;
+use Cline\Struct\Support\CreationContext;
+use Illuminate\Support\Collection;
+
+/**
+ * @author Brian Faust <brian@cline.sh>
+ * @psalm-immutable
+ */
+#[Attribute(Attribute::TARGET_PROPERTY)]
+final readonly class DoesntContainStrict extends AbstractCollectionResultAttribute
+{
+    public function __construct(
+        string $source,
+        public mixed $value,
+        public ?string $key = null,
+    ) {
+        parent::__construct($source);
+    }
+
+    public function computeResult(Collection $items, array $properties, ?CreationContext $context = null): mixed
+    {
+        if ($this->key !== null) {
+            return !$items->containsStrict($this->key, $this->value);
+        }
+
+        return !$items->containsStrict($this->value);
+    }
+}