+++ src/Attributes/CollectionResults/Pull.php
@@ -1,0 +1,37 @@
+<?php declare(strict_types=1);
+
+/**
+ * Copyright (C) Brian Faust
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cline\Struct\Attributes\CollectionResults;
+
+use Attribute;
+use Cline\Struct\Support\CreationContext;
+use Illuminate\Support\Collection;
+
+/**
+ * @author Brian Faust <brian@cline.sh>
+ * @psalm-immutable
+ */
+#[Attribute(Attribute::TARGET_PROPERTY)]
+final readonly class Pull extends AbstractCollectionResultAttribute
+{
+    public function __construct(
+        string $source,
+        public int|string $key,
+        public mixed $default = null,
+    ) {
+        parent::__construct($source);
+    }
+
+    public function computeResult(Collection $items, array $properties, ?CreationContext $context = null): mixed
+    {
+        $items = clone $items;
+
+        return $items->pull($this->key, $this->default);
+    }
+}