+++ src/Attributes/CollectionResults/Reduce.php
@@ -1,0 +1,45 @@
+<?php declare(strict_types=1);
+
+/**
+ * Copyright (C) Brian Faust
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cline\Struct\Attributes\CollectionResults;
+
+use Attribute;
+use Cline\Struct\Contracts\ReducesCollectionItemsInterface;
+use Cline\Struct\Support\CreationContext;
+use Illuminate\Support\Collection;
+
+/**
+ * @author Brian Faust <brian@cline.sh>
+ * @psalm-immutable
+ */
+#[Attribute(Attribute::TARGET_PROPERTY)]
+final readonly class Reduce extends AbstractCollectionResultAttribute
+{
+    /**
+     * @param class-string $callback
+     */
+    public function __construct(
+        string $source,
+        public string $callback,
+        public mixed $initial,
+    ) {
+        parent::__construct($source);
+    }
+
+    public function computeResult(Collection $items, array $properties, ?CreationContext $context = null): mixed
+    {
+        /** @var ReducesCollectionItemsInterface $callback */
+        $callback = $this->resolveCallback($this->callback, ReducesCollectionItemsInterface::class, $context);
+
+        return $items->reduce(
+            static fn (mixed $carry, mixed $value, int|string $key): mixed => $callback->reduce($carry, $value, $key),
+            $this->initial,
+        );
+    }
+}