+++ src/Attributes/CollectionResults/ReduceSpread.php
@@ -1,0 +1,49 @@
+<?php declare(strict_types=1);
+
+/**
+ * Copyright (C) Brian Faust
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cline\Struct\Attributes\CollectionResults;
+
+use Attribute;
+use Cline\Struct\Contracts\ReducesCollectionItemsSpreadInterface;
+use Cline\Struct\Support\CreationContext;
+use Illuminate\Support\Collection;
+
+/**
+ * @author Brian Faust <brian@cline.sh>
+ * @psalm-immutable
+ */
+#[Attribute(Attribute::TARGET_PROPERTY)]
+final readonly class ReduceSpread extends AbstractCollectionResultAttribute
+{
+    /**
+     * @param class-string      $callback
+     * @param array<int, mixed> $initial
+     */
+    public function __construct(
+        string $source,
+        public string $callback,
+        public array $initial,
+    ) {
+        parent::__construct($source);
+    }
+
+    public function computeResult(Collection $items, array $properties, ?CreationContext $context = null): mixed
+    {
+        /** @var ReducesCollectionItemsSpreadInterface $callback */
+        $callback = $this->resolveCallback($this->callback, ReducesCollectionItemsSpreadInterface::class, $context);
+
+        $carry = $this->initial;
+
+        foreach ($items as $value) {
+            $carry = $callback->reduce($carry, $value);
+        }
+
+        return $carry;
+    }
+}