+++ src/Attributes/CollectionSources/AbstractCollectionSourceAttribute.php
@@ -1,0 +1,18 @@
+<?php declare(strict_types=1);
+
+/**
+ * Copyright (C) Brian Faust
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cline\Struct\Attributes\CollectionSources;
+
+use Cline\Struct\Contracts\GeneratesCollectionValueInterface;
+
+/**
+ * @author Brian Faust <brian@cline.sh>
+ * @psalm-immutable
+ */
+abstract readonly class AbstractCollectionSourceAttribute implements GeneratesCollectionValueInterface {}