+++ src/Metadata/CollectionItemDescriptor.php
@@ -54,6 +54,8 @@ final readonly class CollectionItemDescriptor
             isSensitive: false,
             isEncrypted: false,
             isComputed: false,
+            hasCollectionResultAttribute: false,
+            hasCollectionSourceAttribute: false,
             isLazy: false,
             computer: null,
             lazyResolver: null,